Snowflake.Packaging: Move publishing utilities to Snowball.Publishing project
