+++ Packaging/Snowball.Publishing/Snowball.Publishing.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{E84493C8-0171-40A8-94C0-BEFA06C9138B}</ProjectGuid>
+    <ProjectGuid>{B152A58A-7D3C-4C54-9BFC-77023180A4D7}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Snowball</RootNamespace>
-    <AssemblyName>Snowball</AssemblyName>
+    <RootNamespace>Snowball.Publishing</RootNamespace>
+    <AssemblyName>Snowball.Publishing</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -30,6 +30,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Microsoft.Web.XmlTransform, Version=2.1.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Microsoft.Web.Xdt.2.1.1\lib\net40\Microsoft.Web.XmlTransform.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="Newtonsoft.Json, Version=8.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Newtonsoft.Json.8.0.2\lib\net45\Newtonsoft.Json.dll</HintPath>
       <Private>True</Private>
@@ -38,14 +42,12 @@
       <HintPath>..\..\packages\NuGet.Core.2.10.1\lib\net40-Client\NuGet.Core.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="Octokit, Version=0.17.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Octokit.0.17.0\lib\net45\Octokit.dll</HintPath>
+    <Reference Include="Octokit, Version=0.18.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Octokit.0.18.0\lib\net45\Octokit.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.IO.Compression" />
-    <Reference Include="System.IO.Compression.FileSystem" />
     <Reference Include="System.Security" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -55,20 +57,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Installation\PackageKeyPair.cs" />
-    <Compile Include="Packaging\Dependency.cs" />
-    <Compile Include="Packaging\Package.cs" />
-    <Compile Include="Packaging\PackageInfo.cs" />
-    <Compile Include="Packaging\Packagers\Packager.cs" />
-    <Compile Include="Packaging\Packagers\PluginPackager.cs" />
-    <Compile Include="Packaging\Packagers\EmulatorAssemblyPackager.cs" />
-    <Compile Include="Packaging\Packagers\ThemePackager.cs" />
+    <Compile Include="NugetWrapper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Installation\LocalRepository.cs" />
-    <Compile Include="Installation\ReleaseInfo.cs" />
+    <Compile Include="Publisher.cs" />
+    <Compile Include="Secure\AccountKeyStore.cs" />
+    <Compile Include="Secure\GithubAccountManager.cs" />
+    <Compile Include="Secure\PackageKeyStore.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Snowball\Snowball.csproj">
+      <Project>{e84493c8-0171-40a8-94c0-befa06c9138b}</Project>
+      <Name>Snowball</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
