+++ Packaging/Snowball.Publishing/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="8.0.2" targetFramework="net452" />
+  <package id="NuGet.Core" version="2.10.1" targetFramework="net452" />
+  <package id="Octokit" version="0.18.0" targetFramework="net452" />
+</packages>
